centralize the forward declarations
